requirements.txt
# praat-parselmouth==0.4.3
# ipywidgets==8.1.0
# huggingface_hub==0.15.1
## tensorflow==2.11.0
# fairseq==0.12.2
# numpy==1.21
# protobuf==3.9.2

azure-cognitiveservices-speech==1.30.0
python-dotenv==1.0.0
fastapi==0.105.0
uvicorn==0.24.0.post1
# openai-whisper==20230314
# ffmpeg-python
# openai==0.27.5